requirements.txt
streamlit==1.40.2
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
requests==2.32.3
pillow==10.4.0
opencv-python-headless==4.10.0.84
tensorflow==2.17.1
geocoder==1.38.1
pytesseract==0.3.13

